requirements.txt
# Core ML
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.12.0
matplotlib>=3.8.0

# Agent orchestration
crewai>=0.28.0
langchain-google-genai>=1.0.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Dashboard
dash>=2.16.0
plotly>=5.20.0

# Reporting
reportlab>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
pathlib2>=2.3.7
